wip: Refactor 'XYZLut', 'LidarScan' and 'ScanBatcher' to support the new lidar packet profiles, low data profile is failing!

+++ ros2_ouster/include/ros2_ouster/client/point.h
@@ -21,7 +21,7 @@ namespace ouster_ros {
     float intensity;
     uint32_t t;
     uint16_t reflectivity;
-    uint8_t ring;
+    uint16_t ring;
     uint16_t ambient;
     uint32_t range;
     EIGEN_MAKE_ALIGNED_OPERATOR_NEW
@@ -34,7 +34,7 @@ POINT_CLOUD_REGISTER_POINT_STRUCT(
   (float, x, x)(float, y, y)(float, z, z)(float, intensity, intensity)
   // use std::uint32_t to avoid conflicting with pcl::uint32_t
     (std::uint32_t, t, t)(std::uint16_t, reflectivity, reflectivity)(
-    std::uint8_t, ring,
+    std::uint16_t, ring,
     ring)(std::uint16_t, ambient, ambient)(std::uint32_t, range, range)
 )
 // clang-format on
